Adding ability to update file without a scratch file.
Rather than writing the escaped data to a file, the escaped
data is appened to a buffer in memory.  If the buffer containing
the escaped data does not match the buffer containing the existing
file contents, then the original file is truncated and the buffer
containing the escaped data is written to disk.