Updating codetagger to remove tailing spaces in tags
If the tag definition contains an empty line and the tag is indented
with white space, the empty lines in the tag definitions would
result in lines with trailing spaces being inserted into the file.
This update uses a margin stripped of trailing white space for empty
lines found in the tag definition.
file does not cause white space followed